refactor[#45] : @/src -> @ 로 변경
