+++ beta-reader/src/shared/ui/Dropdown/ui/Dropdown.tsx
@@ -1,13 +1,14 @@
 import { VariantProps, cva } from "class-variance-authority";
 import { useRef } from "react";
 
-import { DropdownContext, useDropdown, useDropdownContext } from "../lib";
 import {
   DownLargeIcon,
   DownSmallIcon,
   UpLargeIcon,
   UpSmallIcon
-} from "@/src/shared/assets";
+} from "@/shared/assets";
+
+import { DropdownContext, useDropdown, useDropdownContext } from "../lib";
 
 const dropdownVariant = cva(
   "flex items-center justify-between cursor-pointer",
